xform: move scan-related custom functions to scan_funcs.go
This is one step in breaking up the exploration custom functions into
files based on the rules they facilitate, similar to the normalization
custom functions.

Release note: None
